bug: add  static navbar
